fix bug of plugin

+++ src/bevpool_plugin.cu
@@ -94,7 +94,7 @@ int32_t BEVPoolPlugin::getNbOutputs() const noexcept {
  
 DataType BEVPoolPlugin::getOutputDataType(int32_t index, DataType const *inputTypes, 
                                                                 int32_t nbInputs) const noexcept {
-    return DataType::kHALF;
+    return DataType::kFLOAT;
 }
 
 DimsExprs BEVPoolPlugin::getOutputDimensions(int32_t outputIndex, const DimsExprs *inputs, 
@@ -176,6 +176,7 @@ int32_t BEVPoolPlugin::enqueue(const PluginTensorDesc *inputDesc, const PluginTe
     {
     case int(DataType::kFLOAT):
         if(inputDesc[0].type == DataType::kFLOAT){
+            // printf("bevpool : fp32 fp32\n");
             bev_pool_v2_kernel<float, float><<<grid, block, 0, stream>>>(
                                                         channel, 
                                                         n_intervals,
@@ -191,6 +192,7 @@ int32_t BEVPoolPlugin::enqueue(const PluginTensorDesc *inputDesc, const PluginTe
                                                         reinterpret_cast<float *>(outputs[0]));
         }
         else{
+            // printf("bevpool : fp16 fp32\n");
             bev_pool_v2_kernel<__half, float><<<grid, block, 0, stream>>>(
                                                         channel, 
                                                         n_intervals,
@@ -208,6 +210,7 @@ int32_t BEVPoolPlugin::enqueue(const PluginTensorDesc *inputDesc, const PluginTe
         break;
     case int(DataType::kHALF):
         if(inputDesc[0].type == DataType::kFLOAT){
+            // printf("bevpool : fp32 fp16\n");
             bev_pool_v2_kernel<float, __half><<<grid, block, 0, stream>>>(
                                                         channel, 
                                                         n_intervals,
@@ -223,6 +226,7 @@ int32_t BEVPoolPlugin::enqueue(const PluginTensorDesc *inputDesc, const PluginTe
                                                         reinterpret_cast<__half *>(outputs[0]));
         }
         else{
+            // printf("bevpool : fp16 fp16\n");
             bev_pool_v2_kernel<__half, __half><<<grid, block, 0, stream>>>(
                                                         channel, 
                                                         n_intervals,
